+++ lib/a2z/responses/browse_node.rb
@@ -3,11 +3,16 @@ module A2z
     class BrowseNode
       include Helpers
       
-      attr_accessor :id, :name, :ancestors, :children
+      attr_accessor :id, :name, :ancestors, :children, :most_gifted,
+                    :most_wished_for, :new_releases, :top_sellers
       
       def initialize
         @ancestors = []
         @children = []
+        @most_gifted = []
+        @most_wished_for = []
+        @new_releases = []
+        @top_sellers = []
         @root = false
       end
       
@@ -38,6 +43,20 @@ module A2z
             if data['Ancestors'] && data['Ancestors']['BrowseNode']
               ancestors = array_wrap(data['Ancestors']['BrowseNode'])
               browse_node.ancestors = ancestors.collect { |ancestor| BrowseNode.from_response(ancestor) }
+            end
+            
+            if data['TopItemSet']
+              top_item_sets = array_wrap(data['TopItemSet'])
+              top_item_sets.each do |top_item_set|
+                top_items = array_wrap(top_item_set['TopItem']).collect { |top_item| TopItem.from_response(top_item) }
+                
+                case top_item_set['Type']
+                  when 'MostGifted' then browse_node.most_gifted = top_items
+                  when 'MostWishedFor' then browse_node.most_wished_for = top_items
+                  when 'NewReleases' then browse_node.new_releases = top_items
+                  when 'TopSellers' then browse_node.top_sellers = top_items
+                end
+              end
             end
           end