+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## A2z 0.1.2 (Unreleased)
+
+* Add `most_gifted`, `most_wished_for`, `new_releases`, and `top_sellers`
+  accessors on `BrowseNode`, populating them when present in a browse node
+  lookup response. *Matt Huggins*
+
 ## A2z 0.1.1 (Jan 3, 2012)
 
 * Add `offers` and `offer_summary` accessors on `Item`, populating them when